requirements.txt
selenium==4.11.2
pytest==7.4.3
allure-pytest==2.13.2
allure-python-commons==2.13.2